+++ apps/mobile/src/ui/components/index.ts
@@ -11,5 +11,6 @@ export * from "./MoneyValue";
 export * from "./OfflineBanner";
 export * from "./SelectionField";
 export * from "./Skeleton";
+export * from "./SyncPausedBanner";
 export * from "./SyncStatus";
 export * from "./TransactionRow";